# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SMSOne Enterprise API
  description: |-
    Enterprise Spring Modulith template API. Every response uses the unified envelope ({data | errors, meta, links}) with meta.requestId always present; quote the requestId when reporting issues.

    Operations are grouped as `<axis> · <resource>`. The axis is the authority the operation enforces — Platform (a realm role), Organization (a permission inside one tenant), Shared (any authenticated caller) — and it is read from the code that enforces it, so a group name cannot promise access the API will refuse.
  contact:
    name: SMSOne
    email: ayesigapo@gmail.com
  version: "1"
servers:
- url: http://localhost:8080
  description: Local
- url: https://staging-api.smsone.co.ug
  description: Staging
- url: https://api.smsone.co.ug
  description: Production
security:
- bearerAuth: []
- keycloak: []
...
paths:
...
  /api/v1/orgs/{orgId}/audit:
    get:
      tags:
      - Organization · Audit
      summary: Search one organization's audit trail
      description: "Newest first, narrowed by `action` and an ISO-8601 `from`/`to`\
        \ instant window. For a row written inside an impersonation session `actor`\
        \ is the accountable operator and `onBehalfOf` is the identity they wore."
      operationId: forOrg
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: action
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
      - name: to
        in: query
        required: false
        schema:
          type: string
      - name: "page[size]"
        in: query
        description: "Items per page, 1–100. Defaults to 20."
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - name: "page[after]"
        in: query
        description: Opaque keyset cursor taken from a previous response's meta.page.nextCursor.
          Omit for the first page.
        required: false
        schema:
          type: string
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WindowedResultResourceObject"
...
  /api/v1/audit:
    get:
      tags:
      - Platform · Ops
      summary: Search the platform-wide audit trail
      description: "Newest first, narrowed by any combination of `org`, `action` and\
        \ an ISO-8601 `from`/`to` instant window. For a row written inside an impersonation\
        \ session `actor` is the accountable operator and `onBehalfOf` is the identity\
        \ they wore."
      operationId: platform
      parameters:
      - name: org
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: action
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
      - name: to
        in: query
        required: false
        schema:
          type: string
      - name: "page[size]"
        in: query
        description: "Items per page, 1–100. Defaults to 20."
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - name: "page[after]"
        in: query
        description: Opaque keyset cursor taken from a previous response's meta.page.nextCursor.
          Omit for the first page.
        required: false
        schema:
          type: string
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WindowedResultResourceObject"
  /api/v1/analytics/reports:
    get:
      tags:
      - Platform · Ops
      summary: List the available analytics reports
      operationId: catalog
      parameters:
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceObject"
  /api/v1/analytics/reports/{code}:
    get:
      tags:
      - Platform · Ops
      summary: Run an analytics report
      description: "The report's mart is refreshed from Postgres on every call before\
        \ it is queried, so rows are always current and the response time scales with\
        \ the source data."
      operationId: run
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceObject"
...
components:
  schemas:
...
    ResourceObject:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        attributes: {}
...
    PageMeta:
      type: object
      properties:
        size:
          type: integer
          format: int32
        count:
          type: integer
          format: int32
        hasMore:
          type: boolean
        nextCursor:
          type: string
    WindowedResultResourceObject:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResourceObject"
        page:
          $ref: "#/components/schemas/PageMeta"
  securitySchemes:
    bearerAuth:
      type: http
      description: Paste a Keycloak-issued access token.
      scheme: bearer
      bearerFormat: JWT
    keycloak:
      type: oauth2
      description: Run the Keycloak authorization-code flow (PKCE).
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8081/realms/smsone/protocol/openid-connect/auth
          tokenUrl: http://localhost:8081/realms/smsone/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect
            profile: Profile claims